+++ noodles-sam/src/reader/records.rs
@@ -2,13 +2,14 @@ use std::io::{self, BufRead};
 
 use crate::Record;
 
-use super::Reader;
+use super::{record::Fields, Reader};
 
 /// An iterator over records of a SAM reader.
 ///
 /// This is created by calling [`Reader::records`].
 pub struct Records<'a, R> {
     inner: &'a mut Reader<R>,
+    fields: Fields,
     line_buf: String,
 }
 
@@ -16,9 +17,10 @@ impl<'a, R> Records<'a, R>
 where
     R: BufRead,
 {
-    pub(crate) fn new(inner: &'a mut Reader<R>) -> Self {
+    pub(crate) fn new(inner: &'a mut Reader<R>, fields: Fields) -> Self {
         Self {
             inner,
+            fields,
             line_buf: String::new(),
         }
     }
@@ -36,8 +38,7 @@ where
         match self.inner.read_record(&mut self.line_buf) {
             Ok(0) => None,
             Ok(_) => Some(
-                self.line_buf
-                    .parse()
+                Record::parse_with_fields(&self.line_buf, self.fields)
                     .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e)),
             ),
             Err(e) => Some(Err(e)),